fix: tag pagination
- change pagination variables types: Long -> int
- improve parameter name: filterByName -> filter_by_name

+++ backend/services/innotter-service/src/main/java/com/radiuk/innoter_service/service/TagService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface TagService {
 
-    List<TagResponseDto> getTagsWithPaginationAndLimitAndFilter(Long page, Long limit, String filterByName);
+    List<TagResponseDto> getTagsWithPaginationAndLimitAndFilter(int page, int limit, String filterByName);
 
     TagResponseDto createTag(TagRequestDto tagRequestDto);
 }
